SyncSource: detect ambiguous "backend" value
Some backends react to the "backend" aliases
("calendar/todo/memo/addressbook"). If more than one such backend
is enabled and installed, then these aliases become ambiguous.
Used to be the case when Akonadi and EDS were both enabled.

This commit introduces an error message for that case.

In the case of Akonadi and EDS it is (temporarily) avoided
by not choosing Akonadi when only an alias is used.