SyncContext: "before sync" dump improvements
The "before sync" database dump is always done for a single
database. Adapted the INFO message and function prototype accordingly.

Also detect the case when the "before sync" dump is requested at a
time when it already was done and then just skip it - can happen
when running multiple syncs in a single session.