autotools: compile GDBus before libsyncevolution
libsyncevolution depends on GDBus, which therefore must be dealt
with first.

Mostly the dependencies take care of the right compilation order.
Installing the results is different, items of the same kind are
installed in the order in which they are added to their variables,
which in turn is defined by the order in which .am files are
included.

That broke when making libgdbussyncevo a normal lib like
libsyncevolution (see next commit).