command line: fixed --status and --monitor
The implementation was broken when using GDBus GIO (mismatch between
string and D-Bus object type) and probably also with GDBus libdbus
(missing boolean for GetConfig()).

Rewritten using blocking D-Bus calls, which makes the implementation
also a lot shorter and easier to understand.

It would be good to have automated tests for these features. Not done
yet.