client-test: better logging of item count errors
ASSERT_EQUAL of new/modified/deleted/total items in a source
used to log only the raw numbers when the assertion failed.
Now the comparison is done with an int as expected value and
a full list on the other, and when the assertion fails, the
list is included in the error. Makes debugging a lot easier.

This is implemented by extending CppUnit's assertEqual to a
custom type and defining the necessary traits for it. Would also
be useful for "const char" == std::string comparisons.