testing: Memotoo + local time
Memotoo does not store time zone information in its database. When
talking to a client, it can send either UTC or local time. As
discussed on the mailing list ("[SyncEvolution] Memotoo: conversion to
floating time"), Memotoo now sends all time stamps as floating time
in the time zone configured on the server, without specifying that
time zone in the data. The client must be configured to use the same
time zone.

Floating time is preferred over UTC because then Evolution will
use the local time zone definition when computing recurrences. It
would not do that if UTC was specified.

For some internal reasons, previously Memotoo sometimes still
used UTC. Now that is fixed, which required updating some of
the test data.