autotools: special KDE dependencies for syncevolution.org .deb
Compiling on Ubuntu Lucid and finding package dependencies with
dpkg-shlibdeps led to a .deb which wasn't installable on
Debian Testing because the kdelibs5 and kdepimlibs5 are no
longer provided. They were replaced with individual library
packages.

Handle that by overriding the output of dpkg-shlibdeps when
it contains kdelibs5 or kdepimlibs5 and allow the more recent
packages as alternative.