sys.supp: ignore cond valgrind warning in libdb
Seen on Debian Testing, 64 bit mode, in libdb 5.1.29-1.
Further investigation showed that it is probably harmless,
goes away when compiling libdb with initialization of
padding bytes.

Reported to Debian: http://bugs.debian.org/cgi-bin/bugreport.cgi?bug=662917

I've tracked the root cause in __ham_add_el() in src/hash/hash_page.c:

        if (is_databig) {
                doff.type = H_OFFPAGE;
===>            UMRW_SET(doff.unused[0]);
===>            UMRW_SET(doff.unused[1]);
===>            UMRW_SET(doff.unused[2]);

UMRW_SET is a NOP unless --enable-umrw is used during
configure. Compiling with --enable-umrw avoids the valgrind warning.

The unitialized value is used in a comparison of a checksum with 0.
I can convince myself that this is probably okay, because it doesn't
matter whether some of the data the checksum was calculated over
was uninitialized as long as it doesn't change later on.

But nevertheless, having to deal with such random valgrind errors
while debugging other code isn't nice. Please consider using
--enable-umrw when compiling libdb.