D-Bus server: asynchronous method infrastructure
The asynchronous implementation of D-Bus methods and other, internal
asynchronous events depends on solid support for reporting success and
failure back to the originator of the operation. The
dbus-callbacks.h/cpp files introduce that.

Errors reported to the parent are not logged by the
reporter. Otherwise they would be logged twice, because the parent
cannot tell whether the error was already logged.