WebDAV: support SyncML<->WebDAV bridge
A CalDAV/CardDAV source could not be used as data storage in a SyncML
server or for direct syncing with a SyncML peer (phone, server)
because the source dependent on additional information that was taken
from the "target-config" (syncURL, credentials, log level).

Now the source settings are checked first, with the "target-config" as
fallback:
- "database" - must be set to the full URL of the resource,
  use --print-databases or the servers documentation to find it
- "databaseUser" - username on server
- "databasePassword" - corresponding password

The log level for WebDAV operations is taken from the global logging
instance as fallback and thus from the log settings of the running
sync.

The downside of this approach is further duplication of the
credentials. This will be fixed eventually by introducing the concept
of "credential references", but that is harder to implement and will
depend on the "databaseUser/Password" change anyway.