Exception: rethrow internal exceptions after transmitting via D-Bus
The new Exception::tryRethrowDBus() extracts the original internal
exception attributes (class, parameters) from the string that gets
transmitted via D-Bus and throws recognized exceptions. Will be needed
for syncevo-dbus-helper->syncevo-dbus-server error reporting.

tryRethrow() is the version which works on strings without the D-Bus
exception prefix. It's useful for relaying exceptions from inside an
event loop to the caller outside of the loop by storing the
explanation temporarily in a string.