D-Bus server: rewrote delayed deletion
The previous implementation relied on the glib event loop to free the
allocated shared pointer. If the process quit before executing the
idle callback, that memory was listed as "potentially lost" by
valgrind. To avoid that, track all pending deletions in the Server
instance and explicitly free them in the destructor, in addition to
the idle callback.

Because we don't have a common base class for all items which might
have to be tracked like this, boost::bind is used to automatically
hold a copy of the shared pointer. The goal is not to invoke the
resulting functor, only freeing it matters.