logging: avoid empty format string
gcc is unhappy about empty format strings when format checking is
enabled. -Wno-format-zero-length did not help because it is not
supported for C++ ("enabled by default" - then why does the compiler
complain?!). pragmas are ugly and hiding them inside the logging
macro definition didn't work anyway.

Therefore the easiest solution is to use "\n" instead of "" - also
makes it explicit that a line break is meant with the call.