D-Bus Testing: added assertEqualDiff()
Replaces the diffStrings() from the previous commit. Renamed it
because it really is specifying an assertion and moved it to
DBusUtil because it may be useful outside of the command line
testing.

Also made sure that the diff is part of the assertion because then it
will be repeated when listing the failure.