D-Bus Testing: work around Python 2.6 subprocess bug
Explicitly pass an environment when running the command line
tool. Otherwise subprocess.Popen() from Python 2.6 uses not
os.environ (which would be okay) but rather the environment passed to
a previous call to subprocess.Popen() (which will fail if the previous
test ran with an environment which had SYNCEVOLUTION_DEBUG set).