D-Bus Testing: fixed scanFiles()
The function must return all property assignments, including those
which are commented out. The Python version accidentally included some
comment lines which looked like assignments, partly because it ignored
leading spaces, partly because it looked anywhere in the line for an
equal sign.

Fixed by using a regex for the matching.