GTK-2/3 UI: avoid g_strcasecmp
g_strcasecmp() was deprecated and produces warnings with later glib,
which is fatal with -Werror. The method is used to relax an equal
comparison of the config name. Config names are not case sensitive and
SyncEvolution is using a similar "to lower" approach, so relaxing the
comparison is okay, although it is uncertain why exactly it is
necessary and/or whether it catches all cases.

All strings are expected to be UTF-8, but because SyncEvolution only
does primitive lowercasing itself, calling g_ascii_strcasecmp() seems
like a better match than the full UTF-8 aware case handling methods in
glib.