D-Bus testing: improved status/progress monitoring
DBusUtil.checkSync() now includes a pretty-printed dump of the status
and progress events whenever any of the assertions about the events
fails. This is done in a wrapper method because a) including a message
in the assertions themselves overwrites the assertion information when
using Python 2.6 and b) assing these messages would require changing
lots of lines.

The calls to statusChanged and progressChanges are fixed. They passed
all parameters as a single tuple. This affected
TestSessionAPIsReal.progressChanged(), which compared an integer
against a tuple and always got 'True' as result.

To make checkSync() more flexible, the number of expected reports is
now a parameter with default 1.

The path emitting a signal is included as 'path' keyword for the
callbacks - will be useful for command line testing, which otherwise
wouldn't know which session ran the command line.